Extract peer-link wire I/O helpers into wire_io.py
Fourth PR in the peer_link split arc (follows #769 + #770 +
#773). Moves the WS / Noise plumbing helpers and their two
constants byte-for-byte from peer_link/__init__.py into
peer_link/wire_io.py:

* Constants: _HANDSHAKE_READ_TIMEOUT_SECONDS, _PEER_LABEL_MAX_CHARS
* WS / Noise frame I/O: _read_handshake_message,
  _send_bytes_safely, _send_handshake_message, _send_response
* Payload-content helpers: _parse_intent, _parse_json,
  _str_or_empty, _normalize_label

channel.py's lazy from . import _send_bytes_safely becomes
a clean module-level from .wire_io import _send_bytes_safely
now that the target sibling exists. __init__.py drops the
unused esphome_version + _json top-level imports (both only
used by _send_response, now in wire_io).

Test redirects: the four monkey-patches against
peer_link._HANDSHAKE_READ_TIMEOUT_SECONDS follow the symbol
to peer_link.wire_io._HANDSHAKE_READ_TIMEOUT_SECONDS —
patching the package-level binding wouldn't reach the
wire_io-internal lookup. Test imports of _PEER_LABEL_MAX_CHARS /
_normalize_label / _parse_intent / _parse_json /
_read_handshake_message / _send_bytes_safely /
_send_handshake_message / _send_response redirect to the
wire_io submodule. The docstring-pointer comment in
peer_link_client/one_shot.py follows too.

__init__.py: 893 -> 729 lines (-164). wire_io.py: 189 lines.

All 3389 non-e2e tests pass; ruff + mypy clean.
